requirements.txt
gradio
python-docx
docx2pdf
pymupdf
Pillow
imagehash
camelot-py[cv]
rapidfuzz